Home

pulse input DDM

Code for fitting latent drift diffusion models to pulsed input data and neural activity or behavioral observation data.

Using spock

Key functions

Aggregating data from separate recording sessions

Key functions

Fitting a model to choices

Key functions

Fitting a model to neural activity

Some important functions

Functions

pulse_input_DDM.optimize_model — Method.

optimize_model(pz, bias, pz_fit_vec, bias_fit_vec,
    data; dt, n, map_str, x_tol,f_tol,g_tol, iterations)

Optimize parameters specified within fit vectors.

source

Index

pulse_input_DDM.optimize_model

pulse_input_DDM.optimize_model